Update index.js
Corrected a typo I found on line 6. The word styling was spelled wrong :)

Also, it's much better for a developer to view this file without the comments. Go through all your code and remove any console.logs or comments you made. I know it helps you, but as a rule of thumb - only include comments when it's absolutely necessary.